requirements.txt
flask
requests
paho-mqtt
omnilink-lib